Use model object in angular scope
It's good practice to always throw objects into a 'model' object
attached to a scope to make sure that every path to a scope variable in
the angular template contains at least one dot to avoid breaking
angular's live updating mechanism.
See
http://stackoverflow.com/questions/17178943/does-my-ng-model-really-need-to-have-a-dot-to-avoid-child-scope-problems
for a discussion on why (it basically has to with how angular evaluates
scopes on digest).